+++ plugin/buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -1,8 +1,8 @@
 object PluginVersions {
-    val kotlin = "1.3.10"
-    val ktlintPlugin = "6.3.1"
+    val kotlin = "1.3.21"
+    val ktlintPlugin = "7.1.0"
     val gradlePublishPlugin = "0.10.0"
-    val androidPlugin = "3.2.0"
+    val androidPlugin = "3.3.0"
     val semver = "1.1.1"
     val gradleWrapper = "4.10.2"
     val junit = "4.12"